Refactor names, and use Interface in arguments instead of class
