--- Seedshop/ISeedShop.cs
@@ -1,24 +1,0 @@
-﻿using untitledplantgame.Inventory;
-
-namespace untitledplantgame.Seedshop;
-
-public interface ISeedShop
-{
-	public void SetShopContent(ItemStack[] items);
-	
-	public void GenerateRandomShopStock();
-	public ItemStack[] CurrentStock { get; }
-
-	/// <summary>
-	/// Remove one item of that kind off the shop and returns that instance
-	/// </summary>
-	/// <param name="stack"></param>
-	public ItemStack BuyItem(ItemStack item);
-
-	/// <summary>
-	/// Remove one item of that kind off the shop and returns that instance
-	/// </summary>
-	/// <param name="slotIndex"></param>
-	/// <returns></returns>
-	public ItemStack BuyItem(int slotIndex);
-}
